requirements.txt
matplotlib
opencv-python==4.5.5.64
tqdm
pandas
h5py
numpy==1.24.2
scikit-image
pytorch_msssim
timm
# for demo
engineering_notation
psutil
pyserial
pyzmq
thop
IPython
easygui
pypref
chardet

